requirements.txt
notebook
pytest
pylint
ipdb
jupyterlab
tqdm

# API
fastapi
uvicorn

# Machine Learning
### To be changed later
scikit-learn==1.3.1

# Deep Learning
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
#tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
#tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
#tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

# Data visualization
matplotlib
plotly
seaborn

# Data analysis
numpy==1.23.5
pandas==1.5.3


# Google Cloud
google-cloud-storage

# MLOPS
mlflow==2.1.1
prefect==2.14.9

#.env
python-dotenv

# Frontend
streamlit
requests
pillow